Improve test dialog and level loading
